Task Subscription and Billing completed task

+++ Codeinsight.StreamingManagementSystem/BussinessLogic/DTOs/PaymentDto.cs
@@ -1,10 +1,12 @@
+using Codeinsight.StreamingManagementSystem.BusinessLogic.Enums;
+
 namespace Codeinsight.StreamingManagementSystem.BusinessLogic
 {
     public class PaymentDto
     {
         public int Id { get; set; }
         public int SubscriptionId { get; set; }
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime PaymentDate { get; set; }
         public PaymentStatus Status { get; set; }
     }
